# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/orcamentos": {
      "get": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_orcamentos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "rascunho",
                    "enviado",
                    "aprovado",
                    "recusado",
                    "vencido"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrcamentoLeitura"
                  },
                  "title": "Response Listar Api V1 Orcamentos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_orcamentos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrcamentoEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orcamentos/{id_}": {
      "put": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_orcamentos__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrcamentoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_orcamentos__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orcamentos/{id_}/aprovar": {
      "post": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Aprovar",
        "operationId": "aprovar_api_v1_orcamentos__id___aprovar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AprovacaoSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orcamentos/{id_}/documento": {
      "get": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Documento",
        "operationId": "documento_api_v1_orcamentos__id___documento_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projetos": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_projetos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjetoLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Projetos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projetos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_projetos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/projetos/modelo": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Modelo",
        "operationId": "modelo_api_v1_projetos_modelo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "negocio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Negocio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloProjeto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/etapas-padrao": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Etapas Padrao",
        "operationId": "etapas_padrao_api_v1_projetos_etapas_padrao_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inicio"
            }
          },
          {
            "name": "entrega",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entrega"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EtapaSugerida"
                  },
                  "title": "Response Etapas Padrao Api V1 Projetos Etapas Padrao Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/{id_}": {
      "put": {
        "tags": [
          "Projetos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_projetos__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projetos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_projetos__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/{id_}/relatorio": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Relatorio",
        "operationId": "relatorio_api_v1_projetos__id___relatorio_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AprovacaoSaida": {
        "properties": {
          "orcamento": {
            "$ref": "#/components/schemas/OrcamentoLeitura"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          }
        },
        "type": "object",
        "required": [
          "orcamento",
          "negocio_id"
        ],
        "title": "AprovacaoSaida"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModeloProjeto": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "type": "string",
            "format": "date",
            "title": "Inicio"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "pos_entrega": {
            "type": "string",
            "title": "Pos Entrega"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "cliente_id",
          "negocio_id",
          "orcamento_id",
          "responsavel_id",
          "inicio",
          "escopo",
          "pos_entrega"
        ],
        "title": "ModeloProjeto",
        "description": "Valores iniciais para um projeto novo aberto a partir de um negócio ganho."
      },
...
      "OrcamentoAtualizar": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "validade_dias": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Validade Dias",
            "default": 15
          },
          "status": {
            "type": "string",
            "enum": [
              "rascunho",
              "enviado",
              "aprovado",
              "recusado"
            ],
            "title": "Status",
            "default": "rascunho"
          },
          "desconto": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Desconto",
            "default": "0"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "maxLength": 20000,
            "title": "Obs"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/ItemEntrada"
            },
            "type": "array",
            "minItems": 1,
            "title": "Itens"
          },
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cliente_id",
          "data",
          "itens",
          "versao"
        ],
        "title": "OrcamentoAtualizar"
      },
      "OrcamentoEntrada": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "validade_dias": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Validade Dias",
            "default": 15
          },
          "status": {
            "type": "string",
            "enum": [
              "rascunho",
              "enviado",
              "aprovado",
              "recusado"
            ],
            "title": "Status",
            "default": "rascunho"
          },
          "desconto": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Desconto",
            "default": "0"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "maxLength": 20000,
            "title": "Obs"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/ItemEntrada"
            },
            "type": "array",
            "minItems": 1,
            "title": "Itens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cliente_id",
          "data",
          "itens"
        ],
        "title": "OrcamentoEntrada"
      },
      "OrcamentoLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "validade_dias": {
            "type": "integer",
            "title": "Validade Dias"
          },
          "valido_ate": {
            "type": "string",
            "format": "date",
            "title": "Valido Ate"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_exibido": {
            "type": "string",
            "enum": [
              "rascunho",
              "enviado",
              "aprovado",
              "recusado",
              "vencido"
            ],
            "title": "Status Exibido"
          },
          "desconto": {
            "type": "number",
            "title": "Desconto"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "aprovado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aprovado Em"
          },
          "total_projeto": {
            "type": "number",
            "title": "Total Projeto"
          },
          "total_mensal": {
            "type": "number",
            "title": "Total Mensal"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/ItemLeitura"
            },
            "type": "array",
            "title": "Itens"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "numero",
          "cliente_id",
          "cliente_nome",
          "negocio_id",
          "data",
          "validade_dias",
          "valido_ate",
          "status",
          "status_exibido",
          "desconto",
          "obs",
          "aprovado_em",
          "total_projeto",
          "total_mensal",
          "itens"
        ],
        "title": "OrcamentoLeitura"
      },
...
      "ProjetoLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio"
          },
          "entrega": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrega"
          },
          "objetivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "fora_escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fora Escopo"
          },
          "pos_entrega": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos Entrega"
          },
          "progresso": {
            "type": "integer",
            "title": "Progresso"
          },
          "etapas": {
            "items": {
              "$ref": "#/components/schemas/EtapaLeitura"
            },
            "type": "array",
            "title": "Etapas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "titulo",
          "cliente_id",
          "cliente_nome",
          "negocio_id",
          "orcamento_id",
          "status",
          "responsavel_id",
          "inicio",
          "entrega",
          "objetivo",
          "escopo",
          "fora_escopo",
          "pos_entrega",
          "progresso",
          "etapas"
        ],
        "title": "ProjetoLeitura"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}